Move datesAndDays state to App.jsx
