+++ task-craft/src/components/DateComponent.css
@@ -10,7 +10,6 @@
 
 .date-column {
   text-align: center;
-  margin: 5px;
-  padding: 10px;
-  border: 1px solid #ddd;
+  margin: 3px;
+  padding: 5px;
 }
